feat(module): printRandomNumber Module로 만들고, import로 가져와서 사용하기
